publish_github_release: Improve auto detection of "release" vs "prerelease"
This commit improve the logic checking if the HEAD is associated with
a tag or not. It now excludes pre-release tag when checking if HEAD
is associated with a tag.

The test has also been updated to remove the explicitly deletion of local
tag. Since each test represent a logical progression of releases, there is
no need to delete tags between steps. They are deleted when "reset()" is
called.

Update "githubrelease" dependency from 1.5.1 to 1.5.2:

$ git shortlog 1.5.1..1.5.2 --no-merges
Jean-Christophe Fillion-Robin (7):
      scripts/upload-latest-release: Tweak pre-release name
      print_asset_info: Display asset state
      gh_asset_delete: Support deleting asset independently of their state
      gh_asset_upload: Handle invalid release asset with "new" state
      README: Update list of features
      _update_release_sha: Minimize error rate adding delay between each calls
      CHANGE.md: Replace "Next Release" with "1.5.2"