Cast major and minor version to int
When `major` and `minor` are not casted to `int`, the comparison of CMake version  `3.10.1` will detect that `10` is less than `6` and therefore download CMake win32-86 for versions where the first digit of `major` is less than `6` instead of comparing the whole number. The comparison would work again for values greater than 60, and break again above 100 without casting.